Fix 54 review findings across backends: correctness, memory/concurrency, security, build
Cross-cutting fixes from a full-project review, verified with targeted
affected tests (CPU/CUDA/ROCm/OneAPI/Vulkan), plus several issues found
during verification.

Correctness:
- autograd: LinearBackward higher-order (>2D) batch-dim collapse
- attention: causal mask built additively (no 0*-inf NaN)
- nn: EmbeddingBag include_last_offset bag count; orthogonal_ wide matrices;
  module to() no longer clobbers per-submodule train/eval state
- ops: where() promotes x/y dtypes; broadcast_shapes size-0 vs size-1
- core: unsqueeze stride preserves contiguity; TensorImpl copy ctor no longer
  propagates stale lazy caches to views
- quantization: asymmetric scale for constant tensors (range includes 0)
- jit: Transpose/Squeeze shape inference normalize negative dims
- var/std return NaN when N<=correction on ALL backends (CPU/CUDA/ROCm/
  OneAPI/Vulkan) matching PyTorch
- grid_sample reflection: align_corners-aware forward and +/-1 fold-sign in
  grad_grid, consistent across all 5 backends
- CPU depthwise conv1d/3d, OneAPI std/var/norm/prod: materialize contiguous
  inputs (no more shape-derived reads of non-contiguous views)
- dropout honors manual_seed under OpenMP (counter-based per-element RNG)
- ROCm: native Float64 roi_align (templated); native triangular-solve compiles

Memory/concurrency:
- Vulkan col2im_f64 (Fold) output-centric gather (no float64 RMW race);
  staging-buffer RAII release; descriptor-pool reset when idle
- RAII for GPU scratch buffers on error paths (ROCm linalg/math, OneAPI
  sparse/linalg)
- autograd get_recompute_hooks() returns a locked snapshot copy
- gloo ring_all_reduce: decouple send/recv guards (no deadlock when
  numel<world_size); AVG promotes integers via tenzor::div
- MPI: route element counts through checked_mpi_count; in-place AVG divide
- dispatch enforces same accelerator device index

Security:
- ONNX importer: arity validation, slice/split axis bounds, external_data
  overflow, negative tellg
- TZLITE: bound untrusted element counts before reserve/resize (DoS)
- pickle LONG1 UB shift guard; from_blob deleter acquires the GIL

Build/packaging:
- GPU backends degrade gracefully when toolchain absent (CUDA probe);
  TenzorConfig reports actually-built backends; MPS install/export + glob;
  drop dead ROCm option redefinition

Tests:
- add var/std-NaN and grid_sample-reflection parity regression tests;
  fix GridSample_Bilinear parity test that silently skipped on every GPU

+++ include/tenzor/ops/reduction.hpp
@@ -82,7 +82,14 @@ auto prod(const Tensor& input,
 
 /**
  * @brief Standard deviation of tensor elements.
- * @param unbiased Use unbiased estimator (N-1 denominator)
+ * @param unbiased Use unbiased estimator (N-1 denominator).
+ *
+ * Note on the API: std/var intentionally expose a boolean @p unbiased, while
+ * nanstd/nanvar expose an integer @p correction (Bessel correction). The two
+ * map directly: `unbiased=true` ≡ `correction=1` (N-1 denominator) and
+ * `unbiased=false` ≡ `correction=0` (N denominator). The boolean form is kept
+ * here for source compatibility; use nan* if you need a correction other
+ * than 0 or 1.
  */
 auto std(const Tensor& input,
         std::optional<int64_t> dim = std::nullopt,
@@ -103,7 +110,7 @@ auto var(const Tensor& input,
  * @param p Norm order (1=L1, 2=L2, inf=max)
  */
 auto norm(const Tensor& input,
-         float p = 2.0f,
+         double p = 2.0,
          std::optional<int64_t> dim = std::nullopt,
          bool keepdim = false) -> Tensor;
 
@@ -213,7 +220,7 @@ auto var_mean(const Tensor& input, std::optional<int64_t> dim = std::nullopt,
               bool keepdim = false, bool unbiased = true) -> std::pair<Tensor, Tensor>;
 
 /// p-norm of (a - b)
-auto dist(const Tensor& a, const Tensor& b, float p = 2.0f) -> Tensor;
+auto dist(const Tensor& a, const Tensor& b, double p = 2.0) -> Tensor;
 
 // =========================================================================
 // New reduction operations for PyTorch parity
